benchmark checks correctness
