+++ src/test/scala/edu/berkeley/cs/rise/opaque/OpaqueTestsBase.scala
@@ -23,15 +23,14 @@ import scala.collection.mutable
 import org.apache.log4j.Level
 import org.apache.log4j.LogManager
 import org.apache.spark.sql.SparkSession
-import org.scalactic.TolerantNumerics
-import org.scalactic.Equality
 import org.scalatest.FunSuite
 import org.scalatest.BeforeAndAfterAll
+import org.scalactic.Equality
 import org.scalatest.Tag
 
 import edu.berkeley.cs.rise.opaque.benchmark._
 
-trait OpaqueTestsBase extends FunSuite with BeforeAndAfterAll { self =>
+trait OpaqueTestsBase extends FunSuite with BeforeAndAfterAll with OpaqueTolerance { self =>
 
   def spark: SparkSession
   def numPartitions: Int
@@ -42,25 +41,6 @@ trait OpaqueTestsBase extends FunSuite with BeforeAndAfterAll { self =>
 
   override def afterAll(): Unit = {
     spark.stop()
-  }
-
-  // Modify the behavior of === for Double and Array[Double] to use a numeric tolerance
-  implicit val tolerantDoubleEquality = TolerantNumerics.tolerantDoubleEquality(1e-6)
-
-  def equalityToArrayEquality[A : Equality](): Equality[Array[A]] = {
-    new Equality[Array[A]] {
-      def areEqual(a: Array[A], b: Any): Boolean = {
-        b match {
-          case b: Array[_] =>
-            (a.length == b.length
-              && a.zip(b).forall {
-                case (x, y) => implicitly[Equality[A]].areEqual(x, y)
-              })
-          case _ => false
-        }
-      }
-      override def toString: String = s"TolerantArrayEquality"
-    }
   }
 
   def testAgainstSpark[A : Equality](name: String, testFunc: (String, Tag*) => ((=> Any) => Unit) = test)
