+++ src/test/scala/edu/berkeley/cs/rise/opaque/benchmark/TPCH.scala
@@ -29,7 +29,10 @@ import edu.berkeley.cs.rise.opaque.Utils
 object TPCH {
 
   // Add query numbers here once they are supported
-  val supportedQueries = Seq(1)
+  val supportedQueries = Seq(1, 9)
+
+  // Queries whose results are not defined in a specific order
+  val unorderedQueries = Seq(6, 14, 17, 19)
 
   val tableNames = Seq("part", "supplier", "lineitem", "partsupp", "orders", "nation", "region", "customer")
